% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atbash_cipher.R
\name{atbash_cipher}
\alias{atbash_cipher}
\title{Atbash Cipher}
\usage{
atbash_cipher(text)
}
\arguments{
\item{text}{A character string to be encrypted or decrypted.}
}
\value{
A character string.
}
\description{
A type of substitution cipher where the alphabet is reversed.
}
\examples{
atbash_cipher("Hebrew alphabet")
atbash_cipher("svyivd zokszyvg")
}
