% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar_cipher.R
\name{caesar_cipher}
\alias{caesar_cipher}
\title{Caesar Cipher}
\usage{
caesar_cipher(text, shift)
}
\arguments{
\item{text}{A character string that you want to encrypt or decrypt.}

\item{shift}{An integer specifying the number of positions to shift each letter in the alphabet.}
}
\value{
A character string where each letter in the input text has been shifted by the specified number of positions.
}
\description{
Shift each letter in a string by a certain number of places in the alphabet.
}
\examples{
caesar_cipher("attack at dawn", 3)
caesar_cipher("12345", 5)
caesar_cipher("abcde", -2)
}
