% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse_code.R
\name{morse_code}
\alias{morse_code}
\title{Morse Code}
\usage{
morse_code(text)
}
\arguments{
\item{text}{A character string to be converted to Morse code. The string can include letters, numbers, and spaces.}
}
\value{
A character string representing the input text converted into Morse code. Each letter or number is encoded as a sequence of dots (.) and dashes (-). Words are separated by slashes (/).
}
\description{
Encode messages using Morse code.
}
\examples{
morse_code("Holy biscuits batman")
morse_code("12345")
}
