% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polybius_square.R
\name{polybius_square}
\alias{polybius_square}
\title{Polybius Square}
\usage{
polybius_square(input)
}
\arguments{
\item{input}{A character string to be encrypted or decrypted.}
}
\value{
A character string. (i/j is used interchangeably).
}
\description{
A cipher where each letter is represented by its coordinates in a 5x5 grid.
}
\examples{
polybius_square("Hello friends")
polybius_square("3511331311251543")
polybius_square("24 45 33 24 34 42")
}
