% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenere_cipher.R
\name{vigenere_cipher}
\alias{vigenere_cipher}
\title{Vigenere Cipher}
\usage{
vigenere_cipher(text, keyword)
}
\arguments{
\item{text}{A character string to be encrypted.}

\item{keyword}{A character string serving as the keyword for the cipher. The key is used to determine the letter shifts in the Vigenère cipher.}
}
\value{
A character string where each letter in the input text is shifted according to the corresponding letter in the key.
}
\description{
A cipher where each letter in the plaintext is shifted based on a keyword.
}
\examples{
vigenere_cipher("Short cuts make long delays", "ring")
vigenere_cipher("make it happen", "math")
}
